requirements-dev.txt
Flask>=3.0,<4.0
fastapi>=0.112,<1.0
fakeredis>=2.22,<3.0
httpx>=0.27,<2.0
prometheus-client>=0.20,<1.0
pytest-asyncio>=0.23,<2.0
pydantic>=2.8,<3.0
PyYAML>=6.0,<7.0
redis>=5.0,<8.0
build>=1.2,<2.0
